Make dummy behave a bit more like other providers.

+++ lib/poise_service/service_providers/dummy.rb
@@ -57,8 +57,9 @@ module PoiseService
           end
           Chef::Log.debug("[#{new_resource}] Process environment configured")
           IO.write(pid_file, Process.pid)
-          Chef::Log.debug("[#{new_resource}] PID written to #{pid_file}, dropping privs and execing")
+          Chef::Log.debug("[#{new_resource}] PID written to #{pid_file}")
           Process::UID.change_privilege(new_resource.user)
+          Chef::Log.debug("[#{new_resource}] Changed privs to #{new_resource.user}")
           # Split the command so we don't get an extra sh -c.
           Chef::Log.debug("[#{new_resource}] Execing #{new_resource.command}")
           Kernel.exec(*Shellwords.split(new_resource.command))
@@ -71,16 +72,21 @@ module PoiseService
 
       def action_stop
         return unless pid
+        Chef::Log.debug("[#{new_resource}] Stopping with #{new_resource.stop_signal}. Current PID is #{pid.inspect}.")
         Process.kill(new_resource.stop_signal, pid)
         ::File.unlink(pid_file)
       end
 
       def action_restart
+        return if options['never_restart']
         action_stop
         action_start
       end
 
       def action_reload
+        return if options['never_reload']
+        return unless pid
+        Chef::Log.debug("[#{new_resource}] Reloading with #{new_resource.reload_signal}. Current PID is #{pid.inspect}.")
         Process.kill(new_resource.reload_signal, pid)
       end
